Vagent/requirements.txt
python-dotenv>=1.0.0
aiohttp>=3.8.0
aiortc>=1.5.0
websockets>=10.0
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
